Made requested changes: Implemented SwipeListener and updated that listener.
