+++ app/src/main/java/com/example/habitsmasher/ui/dashboard/HabitListFragment.java
@@ -23,6 +23,7 @@ import com.example.habitsmasher.HabitList;
 import com.example.habitsmasher.ListFragment;
 import com.example.habitsmasher.R;
 import com.example.habitsmasher.User;
+import com.example.habitsmasher.listeners.SwipeListener;
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
 import com.google.android.gms.tasks.Task;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
@@ -121,23 +122,8 @@ public class HabitListFragment extends ListFragment<Habit> {
             }
         })
                 .setSwipeOptionViews(R.id.edit_button, R.id.delete_button)
-                .setSwipeable(R.id.habit_view, R.id.swipe_options, new RecyclerTouchListener.OnSwipeOptionsClickListener() {
-                    @Override
-                    public void onSwipeOptionClicked(int viewID, int position) {
-                        // edit and delete functionality below
-                        switch (viewID){
-                            // if edit button clicked
-                            case R.id.edit_button:
-                                openEditDialogBox(position);
-                                break;
-                            // if delete button clicked
-                            case R.id.delete_button:
-                                updateListAfterDelete(position);
-                                break;
-                        }
-
-                    }
-                });
+                .setSwipeable(R.id.habit_view, R.id.swipe_options,
+                        new SwipeListener(_habitItemAdapter, this, _habitList, _user));
         // connect listener to recycler view
         recyclerView.addOnItemTouchListener(touchListener);
     }
@@ -215,16 +201,6 @@ public class HabitListFragment extends ListFragment<Habit> {
         addHabitDialog.show(getFragmentManager(), "AddHabitDialog");
     }
 
-
-    // note: add this to list fragment class once swipe is complete in habit event list
-    protected void openEditDialogBox(int position) {
-        EditHabitFragment editHabitFragment = new EditHabitFragment(position,
-                _habitItemAdapter._snapshots.get(position),
-                this);
-        editHabitFragment.show(getFragmentManager(),
-                "Edit Habit");
-    }
-
     // note: add this to list fragment class once view is implemented for habitevents
     protected void openViewWindowForItem(int position) {
         // Get the selected habit
@@ -256,15 +232,6 @@ public class HabitListFragment extends ListFragment<Habit> {
     public void updateListAfterEdit(Habit editedHabit, int pos) {
         _habitList.editHabitInDatabase(editedHabit, pos, _user.getId());
         _habitItemAdapter.notifyItemChanged(pos);
-    }
-
-    // add to list fragment class once swipe is fixed in habit events
-    public void updateListAfterDelete(int position) {
-        Habit habitToDelete = _habitItemAdapter._snapshots.get(position);
-        _habitList.deleteHabit(getActivity(),
-                _user.getId(),
-                habitToDelete,
-                position);
     }
 
     @NonNull
